+++ src/store/getters.js
@@ -1,0 +1,5 @@
+export default {
+  validCourses: (state) => {
+    return state.coursesList.filter(course => course);
+  }
+};
