+++ src/store/state.js
@@ -1,0 +1,13 @@
+export default {
+  universityList: [],
+  universtitiesShowed: [],
+  notasByCourse: [],  
+  stateList: [],
+  cityList: [],
+  categoryList: [],
+  modalityList: ["Educação à Distância", "Educação Presencial"],
+  yearList: [ 2017 ],
+  coursesList: [],
+  coursesValids: [],
+  coursesToCompare: []
+};
